+++ topi/python/topi/generic/__init__.py
@@ -19,3 +19,4 @@ from .nn import *
 from .injective import *
 from .extern import *
 from .vision import *
+from .sort import *
