+++ topi/python/topi/generic/vision.py
@@ -188,20 +188,3 @@ def schedule_proposal(outs):
       The computation schedule for the op.
     """
     return _default_schedule(outs, False)
-
-@tvm.target.generic_func
-def schedule_argsort(outs):
-    """Schedule for argsort operator.
-
-    Parameters
-    ----------
-    outs: Array of Tensor
-      The indices that would sort an input array along
-      the given axis.
-
-    Returns
-    -------
-    s: Schedule
-      The computation schedule for the op.
-    """
-    return _default_schedule(outs, False)
